factor out AssetKey into its own module
